+++ UltimaSDK/Ultima/SkillGroups.cs
@@ -54,7 +54,7 @@ namespace Ultima
                         for (int i = 0; i < count - 1; ++i)
                         {
                             int strbuild;
-                            fs.Seek((long)(start + (i * strlen)), SeekOrigin.Begin);
+                            fs.Seek(start + (i * strlen), SeekOrigin.Begin);
                             StringBuilder builder2 = new StringBuilder(17);
                             if (unicode)
                             {
@@ -68,7 +68,7 @@ namespace Ultima
                             }
                             List.Add(new SkillGroup(builder2.ToString()));
                         }
-                        fs.Seek((long)(start + ((count - 1) * strlen)), SeekOrigin.Begin);
+                        fs.Seek(start + ((count - 1) * strlen), SeekOrigin.Begin);
                         try
                         {
                             while (bin.BaseStream.Length != bin.BaseStream.Position)
@@ -90,8 +90,8 @@ namespace Ultima
                 using (BinaryWriter bin = new BinaryWriter(fs))
                 {
                     if (unicode)
-                        bin.Write((int)-1);
-                    bin.Write((int)List.Count);
+                        bin.Write(-1);
+                    bin.Write(List.Count);
 
                     foreach (SkillGroup group in List)
                     {
